Use interrupt driven GPIOs to manage rotary encoder
This completely fixes the flakyness with the encoder on my board, it seems
we aren't polling fast enough for the debouncing algorithm to work properly
on all encoders, this fixes it.

The drawback is quite a few interrupts when manipulating the encoder, I
haven't checked whether this can disrupt audio processing.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>