wifi: Rework/improve wifi status
This removes reliance on the patchbox CLI which is very slow and was only
called occasionally.

Instead, we have our own implementation to query the wifi state more
directly (much faster), running from a separate thread as to avoid
affecting the UI responsiveness.

This allows us to keep a "live" status that is properly updated when
things change from outside of pi-stomp control. This also cuts a few
seconds from pi-stomp startup time (patchbox CLI is that slow).

The next step is probably to have our own UI for configuring the wifi
but that is for another day :-) We probably want to do some serious
UI cleanup work before we do any drastic change such as implementing
an encoder-driven keyboard...

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>